requirements.txt
fastapi==0.110.1
pydantic>=2.6.4
email-validator>=2.2.0
pymongo==4.6.3
python-dotenv>=1.0.1
